Migrate openflowplugim-impl tests to uint types
This migrates the entire openflowplugin-impl test suite to use
Uint types.

We borrow EncodeConstants to get at Uint8-specified
protocol numbers, which is probably not what we want, but then
there seems to be variance of opinion whether to use that or
OFConstants. This is fine for this migration, we will address
the OFConstants/EncodeConstants duplicity when we migrate
production code.

JIRA: OPNFLWPLUG-1099

+++ openflowplugin-impl/src/test/java/org/opendaylight/openflowplugin/impl/connection/testutil/MsgGeneratorTestUtils.java
@@ -14,6 +14,7 @@ import org.opendaylight.yang.gen.v1.urn.opendaylight.openflow.protocol.rev130731
 import org.opendaylight.yang.gen.v1.urn.opendaylight.openflow.protocol.rev130731.multipart.reply.multipart.reply.body.MultipartReplyDescCaseBuilder;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.openflow.protocol.rev130731.multipart.reply.multipart.reply.body.multipart.reply.desc._case.MultipartReplyDesc;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.openflow.protocol.rev130731.multipart.reply.multipart.reply.body.multipart.reply.desc._case.MultipartReplyDescBuilder;
+import org.opendaylight.yangtools.yang.common.Uint32;
 
 public final class MsgGeneratorTestUtils {
 
@@ -30,7 +31,7 @@ public final class MsgGeneratorTestUtils {
 
         MultipartReplyMessageBuilder messageBuilder = new MultipartReplyMessageBuilder()
                 .setMultipartReplyBody(replyBody)
-                .setXid(xid)
+                .setXid(Uint32.valueOf(xid))
                 .setFlags(new MultipartRequestFlags(hasNext))
                 .setType(MultipartType.OFPMPDESC);
         return messageBuilder;
